Add files via upload
An agent-based NetLogo simulation of the Axelrod model for cultural diffusion. The simulation allows to explore the evolution of a system on the basis of the cultural characteristics of the agents that compose it. The dynamics of the model is based on two main mechanisms: (1) the agents tend to interact with other agents who have similar cultural characteristics and (2) during the interaction, the agents influence each other in order to become more similar. This dynamic sometimes leads to cultural homogeneity in which all agents have the same cultural characteristics, but at other times it makes culturally distinct regions develop. The model allows to study to what extent the probability of these two results depends on the parameters of the model: the number of agents in the population; the number of cultural characteristics considered; the number of values ​​that each characteristic can assume (also called cultural traits); the radius of interaction.